Fix menu if multi role

+++ src/main/java/com/dashboard/repository/keuangan/MenuRoleRepo.java
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 public interface MenuRoleRepo extends JpaRepository<MenuRole, Integer> {
     @Query(value = "SELECT * FROM Menu_Role WHERE role_id = ?1", nativeQuery = true)
-    List<MenuRole> findByRoleId(Long id);
+    List<MenuRole> findByRoleId(Integer id);
 
     // @Query(value = "delete from Menu_Role where role_id = ?1", nativeQuery = true)
     // Long deleteByRoleId(Integer id);
